+++ jp/admin/includes/functions/preorder_general.php
@@ -689,16 +689,16 @@ function   tep_pre_order_status_change($oID,$status){
 }
 
 function tep_get_pre_orders_products_string($orders, $single = false, $popup = false, $param_str = '') {
+  global $ocertify;
+
   $str = '';
 
 
   $str .= '<table border="0" cellpadding="2" cellspacing="0" class="popup_order_title" width="100%">';
-  if ($popup) {
-    $str .= '<tr>';
-    $str .= '<td width="20">'.tep_image(DIR_WS_IMAGES.'icon_info.gif',IMAGE_ICON_INFO,16,16).'&nbsp;</td>'; 
-    $str .= '<td align="left"><b>['.$orders['orders_id'].']&nbsp;&nbsp;'.tep_datetime_short_torihiki($orders['date_purchased']).'</b></td>'; 
-    $str .= '<td align="right"><a href="javascript:void(0);" onclick="hideOrdersInfo(1);">X</a></td></tr>';
-  }
+  $str .= '<tr>';
+  $str .= '<td width="20">'.tep_image(DIR_WS_IMAGES.'icon_info.gif',IMAGE_ICON_INFO,16,16).'&nbsp;</td>'; 
+  $str .= '<td align="left"><b>['.$orders['orders_id'].']&nbsp;&nbsp;'.tep_datetime_short_torihiki($orders['date_purchased']).'</b></td>'; 
+  $str .= '<td align="right"><a href="javascript:void(0);" onclick="hideOrdersInfo(1);">X</a></td></tr>';
   $str .= '</table>'; 
   $str .= tep_draw_form('preorders', FILENAME_PREORDERS, urldecode($param_str).'oID='.$orders['orders_id'].'&action=deleteconfirm'); 
   $str .= '<table border="0" cellpadding="0" cellspacing="0" class="popup_order_info" width="100%">';
@@ -1089,7 +1089,9 @@ function tep_get_pre_orders_products_string($orders, $single = false, $popup = f
   $str .= '<tr><td width="220">&nbsp;</td><td class="main" style="padding-left:20%">'; 
   $str .= '<div id="order_del">';
   $str .= '<a href="'.tep_href_link(FILENAME_PREORDERS, urldecode($param_str).'oID='.$orders['orders_id'].'&action=edit').'">'.tep_html_element_button(IMAGE_DETAILS).'</a>'; 
-  $str .= '&nbsp;<a href="javascript:void(0);">'.tep_html_element_button(IMAGE_DELETE, 'onclick="delete_preorder_info(\''.$orders['orders_id'].'\', \''.urlencode($param_str).'\')"').'</a>'; 
+  if ($ocertify->npermission == 15) {
+    $str .= '&nbsp;<a href="javascript:void(0);">'.tep_html_element_button(IMAGE_DELETE, 'onclick="delete_preorder_info(\''.$orders['orders_id'].'\', \''.urlencode($param_str).'\')"').'</a>'; 
+  }
   $str .= '</div>';
   $str .= '</td></tr>'; 
   $str .= '</table>';
